requirements.txt
actionlib==1.14.0
angles==1.9.13
anyio==4.2.0
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
autobahn==17.10.1
Automat==0.8.0
Babel==2.14.0
backcall==0.2.0
base58==2.1.1
bcrypt==3.1.7
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.4
bondpy==1.8.6
breezy==3.0.2
Brlapi==0.7.0
cached-property==1.5.1
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==1.0.0
catkin-pkg-modules==1.0.0
cbor==1.0.0
certifi==2019.11.28
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.3
comm==0.2.1
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cryptography==2.8
cupshelpers==1.0
cv-bridge==1.16.2
cycler==0.10.0
Cython==0.29.14
dataclass-wizard==0.22.2
dbus-python==1.2.16
debugpy==1.8.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.6.0
Deprecated==1.2.7
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distro==1.4.0
distro-info==0.23ubuntu1
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
dockertown==0.2.4
docopt==0.6.2
docutils==0.16
dt-authentication-daffy==0.1.16
dt-data-api-daffy==1.2.0
dtproject==1.0.6
duckietown-docker-utils-daffy==6.1.1
duckietown-shell==5.5.13
dulwich==0.19.15
dynamic-reconfigure==1.7.3
ecdsa==0.18.0
empy==3.3.2
entrypoints==0.3
exceptiongroup==1.2.0
executing==2.0.1
fastimport==0.9.8
fastjsonschema==2.19.1
filelock==3.13.1
fsspec==2023.12.2
future==0.18.3
gazebo-plugins==2.9.2
gazebo-ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
gitdb==4.0.11
GitPython==3.1.41
gpg==1.13.1
httplib2==0.14.0
humanfriendly==10.0
hyperlink==19.0.0
idna==2.8
ifaddr==0.2.0
image-geometry==1.16.2
importlib-metadata==7.0.1
importlib-resources==6.1.1
incremental==16.10.1
interactive-markers==1.12.0
ipykernel==6.29.0
ipython==8.12.3
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
json5==0.9.14
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-client==8.6.0
jupyter-console==6.6.3
jupyter-core==5.7.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter-server==2.12.5
jupyter-server-terminals==0.5.1
jupyterlab==4.0.11
jupyterlab-pygments==0.3.0
jupyterlab-server==2.25.2
jupyterlab-widgets==3.0.9
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
laser-geometry==1.6.7
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
lxml==5.1.0
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
MarkupSafe==2.1.4
matplotlib==3.1.2
matplotlib-inline==0.1.6
message-filters==1.16.0
mistune==3.0.2
more-itertools==4.2.0
mpi4py==3.0.3
mpmath==1.3.0
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.6.0
netifaces==0.10.4
networkx==3.1
nose==1.3.7
notebook==7.0.7
notebook-shim==0.2.3
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.0
olefile==0.46
opencv-python==4.9.0.80
overrides==7.6.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.1
paramiko==2.6.0
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pip==20.0.2
pkgutil-resolve-name==1.3.10
platformdirs==4.1.0
progressbar2==4.3.2
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==3.6.1
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycparser==2.21
pycryptodomex==3.6.1
pycups==1.9.73
pydantic==1.10.13
pydot==1.4.1
PyGithub==1.43.7
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36ubuntu1
python-gitlab==2.0.1
python-gnupg==0.4.5
python-json-logger==2.0.7
python-qt-binding==0.4.4
python-snappy==0.5.3
python-utils==3.8.1
PyTrie==0.2
pytz==2023.3.post1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==25.1.2
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.32.1
reportlab==3.5.34
requests==2.31.0
requests-unixsocket==0.2.0
requirements-parser==0.5.0
resource-retriever==1.12.7
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roman==2.0.0
rosbag==1.16.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosdep==0.22.2
rosdep-modules==0.22.2
rosdistro==0.9.0
rosdistro-modules==0.9.0
rosgraph==1.16.0
rosinstall==0.7.8
rosinstall-generator==0.1.23
roslaunch==1.16.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.16.0
rosmake==1.15.8
rosmaster==1.16.0
rosmsg==1.16.0
rosnode==1.16.0
rosparam==1.16.0
rospkg==1.5.0
rospkg-modules==1.5.0
rospy==1.16.0
rosservice==1.16.0
rostest==1.16.0
rostopic==1.16.0
rosunit==1.15.8
roswtf==1.16.0
rpds-py==0.17.1
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.12
rqt-dep==0.4.12
rqt-graph==0.4.14
rqt-gui==0.5.3
rqt-gui-py==0.5.3
rqt-image-view==0.4.17
rqt-launch==0.4.9
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-msg==0.4.10
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.11
rqt-publisher==0.4.10
rqt-py-common==0.5.3
rqt-py-console==0.4.10
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-robot-steering==0.5.12
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.0
rqt-service-caller==0.4.10
rqt-shell==0.4.11
rqt-srv==0.4.9
rqt-tf-tree==0.6.4
rqt-top==0.4.10
rqt-topic==0.4.13
rqt-web==0.4.10
rviz==1.14.20
scikit-learn==1.3.2
scipy==1.10.1
SecretStorage==2.3.1
Send2Trash==1.8.2
sensor-msgs==1.13.1
service-identity==18.1.0
setuptools==45.2.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
smach==2.5.2
smach-ros==2.5.2
smclib==1.8.6
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
systemd-python==234
termcolor==2.4.0
terminado==0.18.0
texttable==1.6.2
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
threadpoolctl==3.2.0
tinycss2==1.2.1
tomli==2.0.1
topic-tools==1.16.0
torch==2.2.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
triton==2.2.0
Twisted==18.9.0
txaio==2.10.0
typer==0.9.0
types-setuptools==69.0.0.20240115
typing-extensions==4.9.0
tzdata==2023.4
u-msgpack-python==2.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
vcstools==0.1.42
wadllib==1.3.3
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==0.53.0
wheel==0.34.2
whichcraft==0.6.1
widgetsnbextension==4.0.9
wrapt==1.11.2
wsaccel==0.6.2
wstool==0.1.18
xacro==1.14.16
xkit==0.0.0
zeroconf==0.131.0
zipp==1.0.0
zope.interface==4.7.1
pytorch-lightning